requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.7
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
plotly==5.17.0
python-dotenv==1.0.0
APScheduler==3.10.4
requests==2.31.0
Werkzeug==2.3.7
kafka-python==2.0.2
faker==19.3.0
schedule==1.2.0
